third_party/blink/web_tests/external/wpt/css/cssom-view/table-scroll-props-expected.txt

This is a testharness.js-based test.
[FAIL] Table with separated border
  assert_equals: Table with separated border scrollWidth expected 26 but got 20
[FAIL] Table with collapsed border
  assert_equals: Table with collapsed border scrollWidth expected 26 but got 22
Harness: the test ran to completion.
